third_party/blink/web_tests/platform/mac/external/wpt/css/cssom/font-family-serialization-001-expected.txt

This is a testharness.js-based test.
[PASS] Serialization of <generic-family>
[FAIL] Serialization of quoted "<generic-family>"
  assert_equals: expected "\"emoji\"" but got "emoji"
[PASS] Serialization of prefixed -webkit-<generic-family>
[FAIL] Serialization of NonGenericFontFamilyName,-webkit-body,-webkit-standard,-webkit-pictograph,BlinkMacSystemFont
  assert_equals: expected "BlinkMacSystemFont" but got "\"system-ui\""
Harness: the test ran to completion.
